\name{cheng}
\docType{data}
\alias{cheng}
\title{Normalized gene expression Of ischemic cardiomyopathy (ICM) from a publication by Cheng et al. }
\description{
This data is a normalized gene expression matrix of 1500 genes * 10 samples of ischemic cardiomyopathy (ICM) In Homo sapiens.
}
\format{An object of class \code{matrix} with 1500 rows and 10 columns.}
\source{
    \url{https://www.sciencedirect.com/science/article/pii/S0010482520303061?via\%3Dihub}
}

\details{
 Each row and columns correspond to a genes and samples respectively. Rownames  shows the corresponding gene Entrez Id.
}
\usage{cheng}

\keyword{datasets}
