\name{cheng}
\docType{data}
\alias{cheng}
\title{Normalized gene expression Of ischemic cardiomyopathy (ICM) from a publication by Cheng et al. }
\format{An object of class \code{matrix} with 1500 rows and 10 columns.}
\source{
\url{https://www.sciencedirect.com/science/article/pii/S0010482520303061?via\%3Dihub}

Each row and columns correspond to a genes and samples respectively. Rownames  shows the corresponding gene Entrez Id.}
\usage{cheng}

\references{
    \href{https://www.sciencedirect.com/science/article/pii/S0010482520303061?via\%3Dihub}{Cheng, Chew Weng and  J.Beech, David and B.Wheatcroft, Stephen (2020) Advantages of CEMiTool for gene co-expression analysis of RNA-seq data}
}
\keyword{datasets}
