\name{cheng}
\docType{data}
\alias{cheng}
\title{Normalized Gene Expression Of ischemic cardiomyopathy (ICM) In Homo sapiens}
\description{
This data is a normalized gene expression matrix of 2436 genes * 57 samples of ischemic cardiomyopathy (ICM) In Homo sapiens.
}
\usage{cheng}
\format{An object of class matrix with 2436 rows * 57 columns, where rows and columns are the genes and samples respectively. Rownames  shows the corresponding gene Entrez Id.}
\source{World Almanac and Book of Facts, 1975, page 406.}
\references{
    \href{https://www.sciencedirect.com/science/article/pii/S0010482520303061?via\%3Dihub}{Cheng, Chew Weng and  J.Beech, David and B.Wheatcroft, Stephen (2020) Advantages of CEMiTool for gene co-expression analysis of RNA-seq data}
}
\keyword{datasets}
