\name{cheng}
\docType{data}
\alias{cheng}
\title{Normalized gene expression Of ischemic cardiomyopathy (ICM) from a publication by Cheng et al. }
\description{
This data is a normalized gene expression matrix of 1500 genes * 10 samples of ischemic cardiomyopathy (ICM) In Homo sapiens.
}
\usage{cheng}
\format{An object of class matrix with 1500 rows * 10 columns, where rows and columns are the genes and samples respectively. Rownames  shows the corresponding gene Entrez Id.}

\references{
    \href{https://www.sciencedirect.com/science/article/pii/S0010482520303061?via\%3Dihub}{Cheng, Chew Weng and  J.Beech, David and B.Wheatcroft, Stephen (2020) Advantages of CEMiTool for gene co-expression analysis of RNA-seq data}
}
\keyword{datasets}
